style: format and clean up code (by IDEA) (2nd)

+++ src/main/java/net/minecraft/client/renderer/chunk/SetVisibility.java
@@ -1,8 +1,8 @@
 package net.minecraft.client.renderer.chunk;
 
-import java.util.Set;
-
 import net.minecraft.util.EnumFacing;
+
+import java.util.Set;
 
 public class SetVisibility {
     private static final int COUNT_FACES = EnumFacing.values().length;
@@ -62,7 +62,7 @@ public class SetVisibility {
     }
 
     private boolean getBit(int p_getBit_1_) {
-        return (this.bits & (long) (1L << p_getBit_1_)) != 0L;
+        return (this.bits & (1L << p_getBit_1_)) != 0L;
     }
 
     private void setBit(int p_setBit_1_, boolean p_setBit_2_) {
