+++ src/main/java/net/minecraft/world/chunk/EmptyChunk.java
@@ -1,10 +1,6 @@
 package net.minecraft.world.chunk;
 
 import com.google.common.base.Predicate;
-
-import java.util.List;
-import java.util.Random;
-
 import net.minecraft.block.Block;
 import net.minecraft.entity.Entity;
 import net.minecraft.init.Blocks;
@@ -13,6 +9,9 @@ import net.minecraft.util.AxisAlignedBB;
 import net.minecraft.util.BlockPos;
 import net.minecraft.world.EnumSkyBlock;
 import net.minecraft.world.World;
+
+import java.util.List;
+import java.util.Random;
 
 public class EmptyChunk extends Chunk {
     public EmptyChunk(World worldIn, int x, int z) {
@@ -102,7 +101,7 @@ public class EmptyChunk extends Chunk {
     }
 
     public Random getRandomWithSeed(long seed) {
-        return new Random(this.getWorld().getSeed() + (long) ((long) this.xPosition * this.xPosition * 4987142) + (long) (this.xPosition * 5947611L) + (long) ((long) this.zPosition * this.zPosition) * 4392871L + (long) (this.zPosition * 389711L) ^ seed);
+        return new Random(this.getWorld().getSeed() + ((long) this.xPosition * this.xPosition * 4987142) + (this.xPosition * 5947611L) + (long) this.zPosition * this.zPosition * 4392871L + (this.zPosition * 389711L) ^ seed);
     }
 
     public boolean isEmpty() {
